Move legacy tokenizer to plugins package
