+++ flanker/addresslib/plugins/aol.py
@@ -19,7 +19,7 @@
 
 '''
 import re
-from flanker.addresslib.tokenizer import TokenStream
+from flanker.addresslib.plugins.tokenizer import TokenStream
 
 ALPHA      = re.compile(r'''
                         [A-Za-z]+
