+++ flanker/addresslib/plugins/gmail.py
@@ -28,8 +28,8 @@
         dot              ->      .
 '''
 import re
-from flanker.addresslib.tokenizer import TokenStream
-from flanker.addresslib.tokenizer import ATOM
+from flanker.addresslib.plugins.tokenizer import TokenStream
+from flanker.addresslib.plugins.tokenizer import ATOM
 
 
 GMAIL_BASE = re.compile(r'''
