+++ flanker/addresslib/plugins/hotmail.py
@@ -30,7 +30,7 @@
 
 '''
 import re
-from flanker.addresslib.tokenizer import TokenStream
+from flanker.addresslib.plugins.tokenizer import TokenStream
 
 HOTMAIL_PREFIX  = re.compile(r'''
                             [A-Za-z0-9]+
