+++ flanker/addresslib/plugins/icloud.py
@@ -33,7 +33,7 @@
 
 '''
 import re
-from flanker.addresslib.tokenizer import TokenStream
+from flanker.addresslib.plugins.tokenizer import TokenStream
 
 ALPHA          = re.compile(r'''
                             [A-Za-z]+
